+++ apps/nextjs/src/components/AppComponents/AdditionalMaterials/StepLayouts/StepOne.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 
 import { getResourceTypes } from "@oakai/additional-materials/src/documents/additionalMaterials/resourceTypes";
 
@@ -19,21 +19,30 @@ import {
 import { docTypeSelector } from "@/stores/resourcesStore/selectors";
 
 import { useDialog } from "../../DialogContext";
+import FormValidationWarning from "../../FormValidationWarning";
 import ResourcesFooter from "../ResourcesFooter";
 import { handleDialogSelection } from "./helpers";
 
-const StepOne = () => {
-  const { setStepNumber, setDocType, setGeneration } = useResourcesActions();
+const StepOne = ({
+  handleCreateSession,
+}: {
+  handleCreateSession: ({ documentType }: { documentType: string }) => void;
+}) => {
+  const { setDocType, setGeneration, setSubject, setTitle, setYear } =
+    useResourcesActions();
   const docType = useResourcesStore(docTypeSelector);
   const error = useResourcesStore((state) => state.error);
-
+  const [showValidationError, setShowValidationError] = useState("");
   const { setDialogWindow } = useDialog();
 
   useEffect(() => {
     // Reset the document type when the component is mounted
     setDocType(null);
     setGeneration(null);
-  }, [setDocType, setGeneration]);
+    setSubject(null);
+    setTitle(null);
+    setYear(null);
+  }, [setDocType, setGeneration, setSubject, setTitle, setYear]);
 
   const resourceTypes = getResourceTypes().filter(
     (resourceType) => resourceType.isAvailable,
@@ -44,15 +53,21 @@ const StepOne = () => {
   return (
     <>
       <OakFlex $gap={"space-between-m"} $flexDirection="column">
-        <OakFlex $flexDirection={"column"}>
-          <OakFlex $mv={"space-between-s"}>
+        {!!showValidationError && (
+          <FormValidationWarning errorMessage={showValidationError} />
+        )}
+        <OakFlex $mv={"space-between-s"}>
+          <OakFlex $flexDirection={"column"}>
             <OakRadioGroup
               name="radio-group"
               onChange={(value) => {
-                setDocType(value.target.value);
+                const selectedDocType = value.target.value;
+                setDocType(selectedDocType);
                 setGeneration(null);
+                setShowValidationError("");
               }}
               $flexDirection="column"
+              $gap={"space-between-m"}
             >
               {resourceTypes.map((resourceType) => (
                 <OakLabel key={resourceType.id}>
@@ -64,7 +79,7 @@ const StepOne = () => {
                     label={
                       <OakFlex
                         $flexDirection="column"
-                        $gap="all-spacing-2"
+                        // $gap="all-spacing-2"
                         $ml="space-between-ssx"
                       >
                         <OakP $font="body-2-bold">
@@ -82,14 +97,27 @@ const StepOne = () => {
       </OakFlex>
 
       <ResourcesFooter>
-        <OakFlex $justifyContent="flex-end" $width={"100%"}>
+        <OakFlex $justifyContent="space-between" $width={"100%"}>
+          <button onClick={() => console.log("Back a step clicked")}>
+            {/* Todo: Link this up to the previous step when for launch */}
+            <OakFlex $alignItems="center" $gap="all-spacing-2">
+              <OakIcon iconName="chevron-left" />
+              Back a step
+            </OakFlex>
+          </button>
           <OakPrimaryButton
-            onClick={() => setStepNumber(1)}
+            onClick={() => {
+              if (!docType) {
+                setShowValidationError("Please select a teaching material.");
+                return;
+              }
+
+              handleCreateSession({ documentType: docType });
+            }}
             iconName="arrow-right"
             isTrailingIcon={true}
-            disabled={!docType}
           >
-            Continue
+            Next, provide lesson details
           </OakPrimaryButton>
         </OakFlex>
       </ResourcesFooter>