+++ apps/nextjs/src/stores/resourcesStore/index.ts
@@ -1,6 +1,10 @@
 import { create } from "zustand";
 
+import type { TrackFns } from "@/components/ContextProviders/AnalyticsProvider";
+
 import { logStoreUpdates } from "../zustandHelpers";
+import { handleAnalytics } from "./actionFunctions/handleAnalytics";
+import { handleCreateMaterialSession } from "./actionFunctions/handleCreateMaterialSession";
 import { handleDownload } from "./actionFunctions/handleDownload";
 import {
   handleResetFormState,
@@ -66,7 +70,7 @@ const DEFAULT_STATE = {
   refinementGenerationHistory: [],
 };
 
-export const createResourcesStore = () => {
+export const createResourcesStore = (track: TrackFns) => {
   const resourcesStore = create<ResourcesState>()((set, get) => ({
     id: null,
     ...DEFAULT_STATE,
@@ -101,6 +105,11 @@ export const createResourcesStore = () => {
       // History management actions
       undoRefinement: handleUndoRefinement(set, get),
 
+      createMaterialSession: handleCreateMaterialSession(set, get),
+
+      // Analytics actions
+      analytics: handleAnalytics(set, get, track),
+
       // Reset store to default state
       resetToDefault: () =>
         set((state) => ({ ...DEFAULT_STATE, id: state.id })),
@@ -109,5 +118,6 @@ export const createResourcesStore = () => {
 
   // Log store updates
   logStoreUpdates(resourcesStore, "additional-materials");
+
   return resourcesStore;
 };