+++ packages/aila/src/lib/agents/agents.ts
@@ -21,7 +21,6 @@ import {
   MisconceptionsSchemaWithoutLength,
   PriorKnowledgeSchema,
   PriorKnowledgeSctrictMax5Schema,
-  QuizSchema,
   QuizSchemaStrictMax6Schema,
   QuizSchemaWithoutLength,
   SubjectSchema,
@@ -36,8 +35,11 @@ import { learningCyclesInstructions } from "./prompts/learningCyclesInstructions
 import { learningOutcomeInstructions } from "./prompts/learningOutcomeInstructions";
 import { misconceptionsInstructions } from "./prompts/misconceptionsInstructions";
 import { priorKnowledgeInstructions } from "./prompts/priorKnowledgeInstructions";
-import { quizInstructions } from "./prompts/quizInstructions";
+import { identity } from "./prompts/shared/identity";
+import { quizQuestionDesignInstructions } from "./prompts/shared/quizQuestionDesignInstructions";
+import { tier2And3VocabularyDefinitions } from "./prompts/shared/tier2And3VocabularyDefinitions";
 import { starterQuizInstructions } from "./prompts/starterQuizInstructions";
+import { subjectInstructions } from "./prompts/subjectInstructions";
 
 export const agentNames = z.enum([
   "title",
@@ -110,7 +112,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   subject: {
     type: "prompt",
     name: "subject",
-    prompt: "Specify the subject for this lesson.",
+    prompt: subjectInstructions(),
     schemaForLLM: SubjectSchema,
     schemaStrict: SubjectSchema,
     extractRagData: (lp) => lp.subject,
@@ -126,7 +128,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   learningOutcome: {
     type: "prompt",
     name: "learningOutcome",
-    prompt: learningOutcomeInstructions,
+    prompt: learningOutcomeInstructions({ identity }),
     schemaForLLM: LearningOutcomeSchema,
     schemaStrict: LearningOutcomeSchemaStrictMax190,
     extractRagData: (lp) => lp.learningOutcome,
@@ -134,7 +136,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   learningCycles: {
     type: "prompt",
     name: "learningCycles",
-    prompt: learningCycleTitlesInstructions,
+    prompt: learningCycleTitlesInstructions({ identity }),
     schemaForLLM: LearningCyclesSchema,
     schemaStrict: LearningCyclesStrictMax3Schema,
     extractRagData: (lp) => JSON.stringify(lp.learningCycles),
@@ -142,7 +144,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   priorKnowledge: {
     type: "prompt",
     name: "priorKnowledge",
-    prompt: priorKnowledgeInstructions,
+    prompt: priorKnowledgeInstructions({ identity }),
     schemaForLLM: PriorKnowledgeSchema,
     schemaStrict: PriorKnowledgeSctrictMax5Schema,
     extractRagData: (lp) => JSON.stringify(lp.priorKnowledge),
@@ -150,7 +152,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   keyLearningPoints: {
     type: "prompt",
     name: "keyLearningPoints",
-    prompt: keyLearningPointsInstructions,
+    prompt: keyLearningPointsInstructions({ identity }),
     schemaForLLM: KeyLearningPointsSchema,
     schemaStrict: KeyLearningPointsStrictMax5Schema,
     extractRagData: (lp) => JSON.stringify(lp.keyLearningPoints),
@@ -158,7 +160,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   misconceptions: {
     type: "prompt",
     name: "misconceptions",
-    prompt: misconceptionsInstructions,
+    prompt: misconceptionsInstructions({ identity }),
     schemaForLLM: MisconceptionsSchemaWithoutLength,
     schemaStrict: MisconceptionsSchema,
     extractRagData: (lp) => JSON.stringify(lp.misconceptions),
@@ -166,7 +168,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   keywords: {
     type: "prompt",
     name: "keywords",
-    prompt: keywordsInstructions,
+    prompt: keywordsInstructions({ identity, tier2And3VocabularyDefinitions }),
     schemaForLLM: KeywordsSchemaWithoutLength,
     schemaStrict: KeywordsSchema,
     extractRagData: (lp) => JSON.stringify(lp.keywords),
@@ -174,7 +176,10 @@ export const agents: Record<AgentName, AgentDefinition> = {
   starterQuiz: {
     type: "prompt",
     name: "starterQuiz",
-    prompt: [starterQuizInstructions, quizInstructions].join(`\n\n`),
+    prompt: starterQuizInstructions({
+      identity,
+      quizQuestionDesignInstructions,
+    }),
     schemaForLLM: QuizSchemaWithoutLength,
     schemaStrict: QuizSchemaStrictMax6Schema,
     extractRagData: (lp) => JSON.stringify(lp.starterQuiz),
@@ -182,7 +187,10 @@ export const agents: Record<AgentName, AgentDefinition> = {
   cycle: {
     type: "prompt",
     name: "cycle",
-    prompt: learningCyclesInstructions,
+    prompt: learningCyclesInstructions({
+      identity,
+      quizQuestionDesignInstructions,
+    }),
     schemaForLLM: CycleSchemaWithoutLength,
     schemaStrict: CycleSchema,
     extractRagData: (lp) => {
@@ -197,7 +205,7 @@ export const agents: Record<AgentName, AgentDefinition> = {
   exitQuiz: {
     type: "prompt",
     name: "exitQuiz",
-    prompt: [exitQuizInstructions, quizInstructions].join(`\n\n`),
+    prompt: exitQuizInstructions({ identity, quizQuestionDesignInstructions }),
     schemaForLLM: QuizSchemaWithoutLength,
     schemaStrict: QuizSchemaStrictMax6Schema,
     extractRagData: (lp) => JSON.stringify(lp.exitQuiz),
@@ -225,7 +233,10 @@ export const agents: Record<AgentName, AgentDefinition> = {
   additionalMaterials: {
     type: "prompt",
     name: "additionalMaterials",
-    prompt: additionalMaterialsInstructions,
+    prompt: additionalMaterialsInstructions({
+      identity,
+      tier2And3VocabularyDefinitions,
+    }),
     schemaForLLM: AdditionalMaterialsSchema,
     schemaStrict: AdditionalMaterialsSchema,
     extractRagData: (lp) => JSON.stringify(lp.additionalMaterials),